chore: 팔레트용 컬러 enum 파일 분리
